+++ packages/fuselage/src/components/Box/layout.stories.tsx
@@ -34,127 +34,71 @@ export default {
   },
 } as ComponentMeta<typeof Box>;
 
-export const Widths: ComponentStory<typeof Box> = () => (
+export const Borders: ComponentStory<typeof Box> = () => (
   <>
-    <Box w='x64' />
-    <Box width='x64' />
-    <Box w='none' minWidth='x64' />
-    <Box w='sw' maxWidth='x64' />
+    <Box border='2px solid currentColor' />
+    <Box borderBlock='2px solid currentColor' />
+    <Box borderBlockStart='2px solid currentColor' />
+    <Box borderBlockEnd='2px solid currentColor' />
+    <Box borderInline='2px solid currentColor' />
+    <Box borderInlineStart='2px solid currentColor' />
+    <Box borderInlineEnd='2px solid currentColor' />
+    <Box borderWidth='x2' />
+    <Box borderBlockWidth='x2' />
+    <Box borderBlockStartWidth='x2' />
+    <Box borderBlockEndWidth='x2' />
+    <Box borderInlineWidth='x2' />
+    <Box borderInlineStartWidth='x2' />
+    <Box borderInlineEndWidth='x2' />
+    <Box borderWidth='x1' borderStyle='dashed' />
+    <Box borderWidth='x1' borderBlockStyle='dashed' />
+    <Box borderWidth='x1' borderBlockStartStyle='dashed' />
+    <Box borderWidth='x1' borderBlockEndStyle='dashed' />
+    <Box borderWidth='x1' borderInlineStyle='dashed' />
+    <Box borderWidth='x1' borderInlineStartStyle='dashed' />
+    <Box borderWidth='x1' borderInlineEndStyle='dashed' />
+    <Box borderWidth='x4' borderColor='primary-500' />
+    <Box borderWidth='x4' borderBlockColor='primary-500' />
+    <Box borderWidth='x4' borderBlockStartColor='primary-500' />
+    <Box borderWidth='x4' borderBlockEndColor='primary-500' />
+    <Box borderWidth='x4' borderInlineColor='primary-500' />
+    <Box borderWidth='x4' borderInlineStartColor='primary-500' />
+    <Box borderWidth='x4' borderInlineEndColor='primary-500' />
   </>
 );
-Widths.decorators = [
+Borders.decorators = [
   (fn: any) => (
     <Box display='flex' flexWrap='wrap' alignItems='center'>
       {flattenChildren(fn().props.children).map((child: any) =>
-        React.cloneElement(child, { bg: 'neutral', h: 'x32', m: 'x4' })
+        React.cloneElement(child, {
+          size: 'x32',
+          m: 'x16',
+          borderColor: 'stroke-dark',
+        })
       )}
     </Box>
   ),
 ];
 
-export const Heights: ComponentStory<typeof Box> = () => (
+export const BorderRadii: ComponentStory<typeof Box> = () => (
   <>
-    <Box h='x64' />
-    <Box height='x64' />
-    <Box h='none' minHeight='x64' />
-    <Box h='sw' maxHeight='x64' />
+    <Box borderRadius='full' />
+    <Box borderStartStartRadius='full' />
+    <Box borderStartEndRadius='full' />
+    <Box borderEndStartRadius='full' />
+    <Box borderEndEndRadius='full' />
   </>
 );
-Heights.decorators = [
+BorderRadii.decorators = [
   (fn: any) => (
     <Box display='flex' flexWrap='wrap' alignItems='center'>
       {flattenChildren(fn().props.children).map((child: any) =>
-        React.cloneElement(child, { bg: 'neutral', w: 'x32', m: 'x4' })
+        React.cloneElement(child, {
+          bg: 'dark',
+          size: 'x32',
+          m: 'x16',
+        })
       )}
-    </Box>
-  ),
-];
-
-export const Sizes: ComponentStory<typeof Box> = () => (
-  <>
-    <Box size='x64' />
-    <Box size='none' minSize='x64' />
-    <Box size={9999} maxSize='x64' />
-  </>
-);
-Sizes.decorators = [
-  (fn: any) => (
-    <Box display='flex' flexWrap='wrap' alignItems='center'>
-      {flattenChildren(fn().props.children).map((child: any) =>
-        React.cloneElement(child, { bg: 'neutral', m: 'x4' })
-      )}
-    </Box>
-  ),
-];
-
-export const Margins: ComponentStory<typeof Box> = () => (
-  <>
-    <Box bg='neutral'>
-      <Box m={16} bg='tint' size='x16' />
-    </Box>
-    <Box margin={16} />
-    <Box mb={16} />
-    <Box marginBlock={16} />
-    <Box mbs={16} />
-    <Box marginBlockStart={16} />
-    <Box mbe={16} />
-    <Box marginBlockEnd={16} />
-    <Box mi={16} />
-    <Box marginInline={16} />
-    <Box mis={16} />
-    <Box marginInlineStart={16} />
-    <Box mie={16} />
-    <Box marginInlineEnd={16} />
-  </>
-);
-Margins.decorators = [
-  (story: any) => (
-    <Box display='flex' flexWrap='wrap' alignItems='center'>
-      {flattenChildren(story().props.children).map((child: any, i) => (
-        <Box key={i} bg='neutral-200' m={16}>
-          {React.cloneElement(
-            child,
-            { bg: 'primary-200' },
-            <Box bg='neutral-500' size='x16' />
-          )}
-        </Box>
-      ))}
-    </Box>
-  ),
-];
-
-export const Paddings: ComponentStory<typeof Box> = () => (
-  <>
-    <Box bg='neutral'>
-      <Box p={16} bg='tint' size='x16' />
-    </Box>
-    <Box padding={16} />
-    <Box pb={16} />
-    <Box paddingBlock={16} />
-    <Box pbs={16} />
-    <Box paddingBlockStart={16} />
-    <Box pbe={16} />
-    <Box paddingBlockEnd={16} />
-    <Box pi={16} />
-    <Box paddingInline={16} />
-    <Box pis={16} />
-    <Box paddingInlineStart={16} />
-    <Box pie={16} />
-    <Box paddingInlineEnd={16} />
-  </>
-);
-Paddings.decorators = [
-  (story: any) => (
-    <Box display='flex' flexWrap='wrap' alignItems='center'>
-      {flattenChildren(story().props.children).map((child: any, i) => (
-        <Box key={i} bg='neutral-200' m={16}>
-          {React.cloneElement(
-            child,
-            { bg: 'primary-200' },
-            <Box bg='neutral-500' size='x16' />
-          )}
-        </Box>
-      ))}
     </Box>
   ),
 ];
@@ -184,63 +128,22 @@ Display.decorators = [
     </Box>
   ),
 ];
-export const TextAlign: ComponentStory<typeof Box> = () => (
-  <Box display='flex' color='default' flexDirection='column'>
-    <Box textAlign='left'>left</Box>
-    <Divider />
-    <Box textAlign='center'>center</Box>
-    <Divider />
-    <Box textAlign='right'>right</Box>
-    <Divider />
-    <Box textAlign='justify'>justify</Box>
-    <Divider />
-  </Box>
-);
 
-export const VerticalAlign: ComponentStory<typeof Box> = () => (
+export const Elevation: ComponentStory<typeof Box> = () => (
   <>
-    <Box verticalAlign='top' />
-    <Box verticalAlign='baseline' />
-    <Box verticalAlign='sub' />
-    <Box verticalAlign='middle' />
-    <Box verticalAlign='bottom' />
-    <Box verticalAlign='-40px' />
-    <Box verticalAlign='40px' />
+    <Box elevation='0' />
+    <Box elevation='1' />
+    <Box elevation='2' />
+    <Box elevation='1nb' />
+    <Box elevation='2nb' />
   </>
 );
-VerticalAlign.decorators = [
-  (fn: any) => (
-    <Box>
-      {flattenChildren(fn().props.children).map((child: any) =>
-        React.cloneElement(child, {
-          display: 'inline',
-          children: child.props.verticalAlign,
-          color: 'default',
-          border: '1px solid',
-          borderColor: 'stroke-dark',
-          borderRadius: 'x4',
-          m: 'x4',
-          p: 'x4',
-        })
-      )}
-    </Box>
-  ),
-];
-
-export const Position: ComponentStory<typeof Box> = () => (
-  <>
-    <Box position='static' />
-    <Box position='relative' />
-    <Box position='absolute' />
-    <Box position='fixed' />
-  </>
-);
-Position.decorators = [
+Elevation.decorators = [
   (fn: any) => (
     <Box display='flex' flexWrap='wrap' alignItems='center'>
       {flattenChildren(fn().props.children).map((child: any) =>
         React.cloneElement(child, {
-          bg: 'neutral',
+          bg: 'light',
           size: 'x32',
           m: 'x16',
         })
@@ -249,24 +152,19 @@ Position.decorators = [
   ),
 ];
 
-export const ZIndex: ComponentStory<typeof Box> = () => (
+export const Heights: ComponentStory<typeof Box> = () => (
   <>
-    <Box zIndex={2} borderColor='stroke-extra-light-highlight' />
-    <Box zIndex={1} borderColor='stroke-highlight' />
-    <Box zIndex={2} borderColor='stroke-extra-light-error' />
-    <Box zIndex={1} borderColor='stroke-error' />
+    <Box h='x64' />
+    <Box height='x64' />
+    <Box h='none' minHeight='x64' />
+    <Box h='sw' maxHeight='x64' />
   </>
 );
-ZIndex.decorators = [
+Heights.decorators = [
   (fn: any) => (
     <Box display='flex' flexWrap='wrap' alignItems='center'>
       {flattenChildren(fn().props.children).map((child: any) =>
-        React.cloneElement(child, {
-          bg: 'neutral',
-          borderWidth: 'x4',
-          size: 'x32',
-          m: 'neg-x2',
-        })
+        React.cloneElement(child, { bg: 'neutral', w: 'x32', m: 'x4' })
       )}
     </Box>
   ),
@@ -310,21 +208,105 @@ Insets.decorators = [
   ),
 ];
 
-export const Elevation: ComponentStory<typeof Box> = () => (
+export const Invisible: ComponentStory<typeof Box> = () => <Box invisible />;
+
+export const Margins: ComponentStory<typeof Box> = () => (
   <>
-    <Box elevation='0' />
-    <Box elevation='1' />
-    <Box elevation='2' />
-    <Box elevation='1nb' />
-    <Box elevation='2nb' />
+    <Box bg='neutral'>
+      <Box m={16} bg='tint' size='x16' />
+    </Box>
+    <Box margin={16} />
+    <Box mb={16} />
+    <Box marginBlock={16} />
+    <Box mbs={16} />
+    <Box marginBlockStart={16} />
+    <Box mbe={16} />
+    <Box marginBlockEnd={16} />
+    <Box mi={16} />
+    <Box marginInline={16} />
+    <Box mis={16} />
+    <Box marginInlineStart={16} />
+    <Box mie={16} />
+    <Box marginInlineEnd={16} />
   </>
 );
-Elevation.decorators = [
+Margins.decorators = [
+  (story: any) => (
+    <Box display='flex' flexWrap='wrap' alignItems='center'>
+      {flattenChildren(story().props.children).map((child: any, i) => (
+        <Box key={i} bg='neutral-200' m={16}>
+          {React.cloneElement(
+            child,
+            { bg: 'primary-200' },
+            <Box bg='neutral-500' size='x16' />
+          )}
+        </Box>
+      ))}
+    </Box>
+  ),
+];
+
+export const Opacity: ComponentStory<typeof Box> = () => (
+  <Box display='flex'>
+    <Box size={32} opacity={0.1} bg='dark' />
+    <Box size={32} opacity={0.3} bg='dark' />
+    <Box size={32} opacity={0.5} bg='dark' />
+    <Box size={32} opacity={0.7} bg='dark' />
+    <Box size={32} opacity={0.9} bg='dark' />
+    <Box size={32} opacity={1} bg='dark' />
+  </Box>
+);
+
+export const Paddings: ComponentStory<typeof Box> = () => (
+  <>
+    <Box bg='neutral'>
+      <Box p={16} bg='tint' size='x16' />
+    </Box>
+    <Box padding={16} />
+    <Box pb={16} />
+    <Box paddingBlock={16} />
+    <Box pbs={16} />
+    <Box paddingBlockStart={16} />
+    <Box pbe={16} />
+    <Box paddingBlockEnd={16} />
+    <Box pi={16} />
+    <Box paddingInline={16} />
+    <Box pis={16} />
+    <Box paddingInlineStart={16} />
+    <Box pie={16} />
+    <Box paddingInlineEnd={16} />
+  </>
+);
+Paddings.decorators = [
+  (story: any) => (
+    <Box display='flex' flexWrap='wrap' alignItems='center'>
+      {flattenChildren(story().props.children).map((child: any, i) => (
+        <Box key={i} bg='neutral-200' m={16}>
+          {React.cloneElement(
+            child,
+            { bg: 'primary-200' },
+            <Box bg='neutral-500' size='x16' />
+          )}
+        </Box>
+      ))}
+    </Box>
+  ),
+];
+
+export const Position: ComponentStory<typeof Box> = () => (
+  <>
+    <Box position='static' />
+    <Box position='relative' />
+    <Box position='absolute' />
+    <Box position='fixed' />
+  </>
+);
+Position.decorators = [
   (fn: any) => (
     <Box display='flex' flexWrap='wrap' alignItems='center'>
       {flattenChildren(fn().props.children).map((child: any) =>
         React.cloneElement(child, {
-          bg: 'light',
+          bg: 'neutral',
           size: 'x32',
           m: 'x16',
         })
@@ -333,4 +315,103 @@ Elevation.decorators = [
   ),
 ];
 
-export const Invisible: ComponentStory<typeof Box> = () => <Box invisible />;
+export const Widths: ComponentStory<typeof Box> = () => (
+  <>
+    <Box w='x64' />
+    <Box width='x64' />
+    <Box w='none' minWidth='x64' />
+    <Box w='sw' maxWidth='x64' />
+  </>
+);
+Widths.decorators = [
+  (fn: any) => (
+    <Box display='flex' flexWrap='wrap' alignItems='center'>
+      {flattenChildren(fn().props.children).map((child: any) =>
+        React.cloneElement(child, { bg: 'neutral', h: 'x32', m: 'x4' })
+      )}
+    </Box>
+  ),
+];
+
+export const Sizes: ComponentStory<typeof Box> = () => (
+  <>
+    <Box size='x64' />
+    <Box size='none' minSize='x64' />
+    <Box size={9999} maxSize='x64' />
+  </>
+);
+Sizes.decorators = [
+  (fn: any) => (
+    <Box display='flex' flexWrap='wrap' alignItems='center'>
+      {flattenChildren(fn().props.children).map((child: any) =>
+        React.cloneElement(child, { bg: 'neutral', m: 'x4' })
+      )}
+    </Box>
+  ),
+];
+
+export const TextAlign: ComponentStory<typeof Box> = () => (
+  <Box display='flex' color='default' flexDirection='column'>
+    <Box textAlign='left'>left</Box>
+    <Divider />
+    <Box textAlign='center'>center</Box>
+    <Divider />
+    <Box textAlign='right'>right</Box>
+    <Divider />
+    <Box textAlign='justify'>justify</Box>
+    <Divider />
+  </Box>
+);
+
+export const VerticalAlign: ComponentStory<typeof Box> = () => (
+  <>
+    <Box verticalAlign='top' />
+    <Box verticalAlign='baseline' />
+    <Box verticalAlign='sub' />
+    <Box verticalAlign='middle' />
+    <Box verticalAlign='bottom' />
+    <Box verticalAlign='-40px' />
+    <Box verticalAlign='40px' />
+  </>
+);
+VerticalAlign.decorators = [
+  (fn: any) => (
+    <Box>
+      {flattenChildren(fn().props.children).map((child: any) =>
+        React.cloneElement(child, {
+          display: 'inline',
+          children: child.props.verticalAlign,
+          color: 'default',
+          border: '1px solid',
+          borderColor: 'stroke-dark',
+          borderRadius: 'x4',
+          m: 'x4',
+          p: 'x4',
+        })
+      )}
+    </Box>
+  ),
+];
+
+export const ZIndex: ComponentStory<typeof Box> = () => (
+  <>
+    <Box zIndex={2} borderColor='stroke-extra-light-highlight' />
+    <Box zIndex={1} borderColor='stroke-highlight' />
+    <Box zIndex={2} borderColor='stroke-extra-light-error' />
+    <Box zIndex={1} borderColor='stroke-error' />
+  </>
+);
+ZIndex.decorators = [
+  (fn: any) => (
+    <Box display='flex' flexWrap='wrap' alignItems='center'>
+      {flattenChildren(fn().props.children).map((child: any) =>
+        React.cloneElement(child, {
+          bg: 'neutral',
+          borderWidth: 'x4',
+          size: 'x32',
+          m: 'neg-x2',
+        })
+      )}
+    </Box>
+  ),
+];
